new module structure under load_balancing
